+++ CHANGELOG.md
@@ -28,9 +28,11 @@ All notable changes to this project will be documented in this file.
 
 - Neetcode 75 question list.
 
-### Fixed
+### Changed
 
 - Not null constraints on the fields that are never null from the server.
+- `QuestionModelContainer { question: RefCell<QuestionModel> }` changed to `Rc<RefCell<QuestionModel>>`
+    - As prior implemented hash. Hashables should not be mutable.
 
 ## [0.2.0] - 2023-07-30